fix errors: add types and generic types
