+++ src/app/components/todos/todo-card/todo-card.component.ts
@@ -1,5 +1,6 @@
 import { NgFor } from '@angular/common';
 import { Component, EventEmitter, Input, input, Output, output } from '@angular/core';
+import { Todo } from '../todo-interface';
 
 @Component({
   selector: 'app-todo-card',
@@ -10,7 +11,7 @@ import { Component, EventEmitter, Input, input, Output, output } from '@angular/
 })
 export class TodoCardComponent {
   @Input()
-  todo: any
+  todo!: Todo
   @Output()
   deleteTodo = new EventEmitter();
