+++ src/app/components/todos/todos.component.ts
@@ -2,13 +2,9 @@ import { NgFor } from "@angular/common";
 import { Component, inject } from "@angular/core";
 import {  TodoCardComponent } from "./todo-card/todo-card.component";
 import { TodosApiService } from "../../todos-api.service";
+import { Todo } from "./todo-interface";
 
-export interface Todo {
-  userId: number;
-  id: number;
-  title: string;
-  completed: boolean;
-}
+
 
 @Component({
   selector: 'app-users-list',
@@ -22,12 +18,12 @@ export class TodosListComponent {
   todos: Todo[] = [];
 
   constructor() {
-    this.todosApiService.getTodos().subscribe((response: any) => {
+    this.todosApiService.getTodos().subscribe((response: Todo[]) => {
       this.todos = response;
     });
   }
 
-  deleteTodo(id: any) {
+  deleteTodo(id: number) {
     this.todos = this.todos.filter(
         todo => {
       if (id === todo.id) {
