+++ src/app/components/users/user-interface.ts
@@ -1,0 +1,23 @@
+export interface User {
+    id: number;
+    name: string;
+    username: string;
+    email: string;
+    adress: {
+      street: string;
+      suite: string;
+      city: string;
+      zipcode: string;
+      geo: {
+        lat: string;
+        lng: string;
+      };
+    };
+    phone: string;
+    website: string;
+    company: {
+      name: string;
+      catchphrase: string;
+      bs: string;
+    };
+  }
