+++ src/app/todos-api.service.ts
@@ -1,11 +1,12 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
+import { Todo } from './components/todos/todo-interface';
 
 @Injectable({ providedIn: 'root' })
 export class TodosApiService {
   readonly apiService = inject(HttpClient);
 
   getTodos() {
-    return this.apiService.get('https://jsonplaceholder.typicode.com/todos');
+    return this.apiService.get<Todo[]>('https://jsonplaceholder.typicode.com/todos');
   }
 }
